@@ -1,6 +1,6 @@
 {
   "name": "@hoprnet/chat-bot",
-  "version": "1.37.0-alpha.0",
+  "version": "1.49.0-alpha.0",
   "description": "",
   "repository": "https://github.com/hoprnet/hoprnet.git",
   "homepage": "https://hoprnet.org",
@@ -17,28 +17,25 @@
     "start": "node dist/index.js",
     "start:basodino": "DEBUG=hopr-chatbot* BOT_NAM=randobot ETHEREUM_PROVIDER=wss://xdai.poanetwork.dev/wss HOST_IPV4=0.0.0.0:9093 HOPR_BOOTSTRAP_ADDRESS=_dnsaddr.bootstrap.basodino.develop.hoprnet.org node dist/index.js",
     "lint": "eslint \"{src,apps,libs,test}/**/*.ts\" --fix",
-    "test": "jest",
-    "test:watch": "jest --watch",
-    "test:cov": "jest --coverage",
-    "test:debug": "node --inspect-brk -r tsconfig-paths/register -r ts-node/register node_modules/.bin/jest --runInBand"
+    "test": "mocha"
   },
   "engines": {
     "node": "12",
     "yarn": "1"
   },
   "dependencies": {
-    "@hoprnet/hopr-core": "1.37.0-alpha.0",
-    "@hoprnet/hopr-core-connector-interface": "1.37.0-alpha.0",
-    "@hoprnet/hopr-core-ethereum": "1.37.0-alpha.0",
-    "@hoprnet/hopr-utils": "1.37.0-alpha.0",
-    "@hoprnet/twitter-api-client": "^0.1.8",
+    "@hoprnet/hopr-core": "1.49.0-alpha.0",
+    "@hoprnet/hopr-core-connector-interface": "1.49.0-alpha.0",
+    "@hoprnet/hopr-core-ethereum": "1.49.0-alpha.0",
+    "@hoprnet/hopr-utils": "1.49.0-alpha.0",
+    "@hoprnet/twitter-api-client": "^0.1.9",
     "@linkdrop/sdk": "1.1.6",
     "debug": "4.2.0",
     "dotenv": "^8.2.0",
     "dotenv-parse-variables": "^1.0.1",
     "firebase-admin": "^9.1.1",
     "global": "^4.4.0",
-    "peer-id": "0.13",
+    "peer-id": "0.14.2",
     "rimraf": "^3.0.2",
     "tinyurl": "1.1.7",
     "wait-for-stuff": "^1.4.0",
@@ -48,33 +45,25 @@
     "@types/dotenv": "^8.2.0",
     "@types/dotenv-parse-variables": "^0.2.0",
     "@types/express": "^4.17.3",
-    "@types/jest": "26.0.15",
+    "@types/mocha": "^8.0.3",
     "@types/node": "~12",
     "@typescript-eslint/eslint-plugin": "4.0.0",
     "@typescript-eslint/parser": "3.10.1",
-    "eslint": "7.11.0",
+    "eslint": "7.12.1",
     "eslint-config-prettier": "^6.10.0",
     "eslint-plugin-import": "^2.20.1",
-    "jest": "26.6.1",
+    "mocha": "^8.0.1",
     "prettier": "^2.0.5",
-    "ts-jest": "26.1.0",
     "ts-loader": "^8.0.4",
     "ts-node": "^9.0.0",
     "tsconfig-paths": "^3.9.0",
     "typescript": "^4.0.3"
   },
-  "jest": {
-    "moduleFileExtensions": [
-      "js",
-      "json",
+  "mocha": {
+    "extension": [
       "ts"
     ],
-    "rootDir": "src",
-    "testRegex": ".spec.ts$",
-    "transform": {
-      "^.+\\.(t|j)s$": "ts-jest"
-    },
-    "coverageDirectory": "coverage",
-    "testEnvironment": "node"
+    "spec": "src/**/*.spec.ts",
+    "require": "ts-node/register"
   }
 }
