@@ -8,82 +8,79 @@ import SECIO = require('libp2p-secio')
 // @ts-ignore
 import TCP from 'libp2p-tcp'
 
-import { Handler, MultiaddrConnection } from '../../@types/transport'
-
+import { Handler, MultiaddrConnection } from 'libp2p'
 import Multiaddr from 'multiaddr'
-import PeerInfo from 'peer-info'
 import pipe from 'it-pipe'
-
 import Relay from './relay'
 import { randomBytes } from 'crypto'
-
 import { privKeyToPeerId } from '../../utils'
-import { WebRTCUpgrader } from './webrtc'
+import { getAddress } from '../../test-utils'
+import assert from 'assert'
 
 const TEST_PROTOCOL = `/test/0.0.1`
 
 const privKeys = [randomBytes(32), randomBytes(32), randomBytes(32)]
 
-describe('should create a socket and connect to it', function () {
-  async function generateNode(options: {
-    id: number
-    ipv4?: boolean
-    ipv6?: boolean
-    connHandler?: (conn: MultiaddrConnection) => void
-  }): Promise<libp2p> {
-    const peerInfo = new PeerInfo(await privKeyToPeerId(privKeys[options.id]))
+async function generateNode(options: {
+  id: number
+  ipv4?: boolean
+  ipv6?: boolean
+  connHandler?: (conn: MultiaddrConnection) => void
+}): Promise<{ node: libp2p; relay: Relay }> {
+  const peerId = await privKeyToPeerId(privKeys[options.id])
+  let addresses = []
 
-    if (options.ipv4) {
-      peerInfo.multiaddrs.add(
-        Multiaddr(`/ip4/127.0.0.1/tcp/${9490 + 2 * options.id}`).encapsulate(`/p2p/${peerInfo.id.toB58String()}`)
-      )
-    }
-
-    if (options.ipv6) {
-      peerInfo.multiaddrs.add(
-        Multiaddr(`/ip6/::1/tcp/${9490 + 2 * options.id + 1}`).encapsulate(`/p2p/${peerInfo.id.toB58String()}`)
-      )
-    }
-
-    const node = new libp2p({
-      peerInfo,
-      modules: {
-        transport: [TCP],
-        streamMuxer: [MPLEX],
-        connEncryption: [SECIO],
-        dht: KadDHT
-      },
-      config: {
-        dht: {
-          enabled: false
-        },
-        relay: {
-          enabled: false
-        },
-        peerDiscovery: {
-          autoDial: false
-        }
-      }
-    })
-
-    //@ts-ignore
-    node.relay = new Relay(node, options.connHandler)
-
-    node.handle([TEST_PROTOCOL], (handler: Handler) => {
-      pipe(
-        handler.stream,
-        // echoing msg
-        handler.stream
-      )
-    })
-
-    await node.start()
-
-    return node
+  if (options.ipv4) {
+    addresses = [Multiaddr(`/ip4/127.0.0.1/tcp/${9490 + 2 * options.id}`).encapsulate(`/p2p/${peerId.toB58String()}`)]
   }
 
+  if (options.ipv6) {
+    addresses = [Multiaddr(`/ip6/::1/tcp/${9490 + 2 * options.id + 1}`).encapsulate(`/p2p/${peerId.toB58String()}`)]
+  }
+
+  const node = new libp2p({
+    peerId,
+    addresses: { listen: addresses },
+    modules: {
+      transport: [TCP],
+      streamMuxer: [MPLEX],
+      connEncryption: [SECIO],
+      dht: KadDHT
+    },
+    config: {
+      dht: {
+        enabled: false
+      },
+      relay: {
+        enabled: false
+      },
+      peerDiscovery: {
+        autoDial: false
+      }
+    }
+  })
+
+  let relay = new Relay(node, options.connHandler)
+
+  node.handle([TEST_PROTOCOL], (handler: Handler) => {
+    pipe(
+      handler.stream,
+      // echoing msg
+      handler.stream
+    )
+  })
+
+  await node.start()
+
+  return {
+    node,
+    relay
+  }
+}
+
+describe('transport/relay.spec.ts should create a socket and connect to it', function () {
   it('should create a node and echo a single message', async function () {
-    let [sender, relay, counterparty] = await Promise.all([
+    let [sender, relayer, counterparty] = await Promise.all([
       generateNode({ id: 0, ipv4: true }),
       generateNode({ id: 1, ipv4: true }),
       generateNode({
@@ -95,7 +92,7 @@ describe('should create a socket and connect to it', function () {
             (source: AsyncIterable<Uint8Array>) => {
               return (async function* () {
                 for await (const msg of source) {
-                  console.log(`echoing <${new TextDecoder().decode(msg.slice())}>`)
+                  //console.log(`echoing <${new TextDecoder().decode(msg.slice())}>`)
                   yield msg
                 }
               })()
@@ -107,12 +104,11 @@ describe('should create a socket and connect to it', function () {
     ])
 
     // Make sure that the nodes know each other
-    await Promise.all([sender.dial(relay.peerInfo), counterparty.dial(relay.peerInfo)])
-    //@ts-ignore
+    await Promise.all([sender.node.dial(getAddress(relayer.node)), counterparty.node.dial(getAddress(relayer.node))])
+
     let conn = await sender.relay.establishRelayedConnection(
-      Multiaddr(`/p2p/${counterparty.peerInfo.id.toB58String()}`),
-      [relay.peerInfo],
-      new WebRTCUpgrader({})
+      Multiaddr(`/p2p/${counterparty.node.peerId.toB58String()}`),
+      relayer.node.multiaddrs
     )
 
     await pipe(
@@ -126,13 +122,13 @@ describe('should create a socket and connect to it', function () {
       conn,
       async (source: AsyncIterable<Uint8Array>) => {
         setTimeout(() => setImmediate(() => conn.close()), 300)
-        for await (const msg of source) {
-          console.log(`received <${new TextDecoder().decode(msg.slice())}>`)
+        for await (const _msg of source) {
+          //console.log(`received <${new TextDecoder().decode(msg.slice())}>`)
         }
       }
     )
 
-    await counterparty.stop()
+    await counterparty.node.stop()
 
     await new Promise((resolve) => setTimeout(resolve, 200))
 
@@ -145,7 +141,7 @@ describe('should create a socket and connect to it', function () {
           (source: AsyncIterable<Uint8Array>) => {
             return (async function* () {
               for await (const msg of source) {
-                console.log(`echoing <${new TextDecoder().decode(msg.slice())}>`)
+                //console.log(`echoing <${new TextDecoder().decode(msg.slice())}>`)
                 yield msg
               }
             })()
@@ -155,20 +151,16 @@ describe('should create a socket and connect to it', function () {
       }
     })
 
-    await counterparty.dial(relay.peerInfo)
-    //@ts-ignore
+    await counterparty.node.dial(getAddress(relayer.node))
     conn = await sender.relay.establishRelayedConnection(
-      Multiaddr(`/p2p/${counterparty.peerInfo.id.toB58String()}`),
-      [relay.peerInfo],
-      new WebRTCUpgrader({})
+      Multiaddr(`/p2p/${counterparty.node.peerId.toB58String()}`),
+      relayer.node.multiaddrs
     )
 
     await pipe(
       (async function* () {
         yield new TextEncoder().encode(`first message`)
-
         await new Promise((resolve) => setTimeout(resolve, 100))
-
         yield new TextEncoder().encode(`second message`)
       })(),
       conn,
@@ -180,6 +172,7 @@ describe('should create a socket and connect to it', function () {
       }
     )
 
+    await Promise.all([sender, relayer, counterparty].map((x) => x.node.stop()))
     await new Promise((resolve) => setTimeout(resolve, 200))
   })
 
@@ -439,18 +432,19 @@ describe('should create a socket and connect to it', function () {
   })
 
   it('should not use itself as relay node', async function () {
-    // let [sender, counterparty] = await Promise.all([
-    //   generateNode({ id: 0, ipv4: true }),
-    //   generateNode({ id: 2, ipv4: true }),
-    // ])
-    // let errThrown = false
-    // try {
-    //   await sender.relay.establishRelayedConnection(Multiaddr(`/p2p/${counterparty.peerInfo.id.toB58String()}`), [
-    //     sender.peerInfo,
-    //   ])
-    // } catch (err) {
-    //   errThrown = true
-    // }
-    // assert(errThrown, `Must throw an error if there is no other opportunity than calling ourself`)
+    let [sender, counterparty] = await Promise.all([
+      generateNode({ id: 0, ipv4: true }),
+      generateNode({ id: 2, ipv4: true })
+    ])
+    let errThrown = false
+    try {
+      await sender.relay.establishRelayedConnection(
+        Multiaddr(`/p2p/${counterparty.node.peerId.toB58String()}`),
+        sender.node.multiaddrs
+      )
+    } catch (err) {
+      errThrown = true
+    }
+    assert(errThrown, `Must throw an error if there is no other opportunity than calling ourself`)
   })
 })
